apdu: limit LC to 255 bytes and update docs

+++ src/apdu-pubkey.c
@@ -122,11 +122,11 @@ encode_xpub(
   return b58enc(b58, b58_sz, data, sizeof(data));
 }
 
-volatile uint16_t
+volatile uint8_t
 hns_apdu_get_public_key(
   uint8_t p1,
   uint8_t p2,
-  uint16_t len,
+  uint8_t len,
   volatile uint8_t *buf,
   volatile uint8_t *out,
   volatile uint8_t *flags
