+++ src/apdu-version.c
@@ -6,11 +6,11 @@
 #include "apdu.h"
 #include "ledger.h"
 
-volatile uint16_t
+volatile uint8_t
 hns_apdu_get_app_version(
   uint8_t p1,
   uint8_t p2,
-  uint16_t len,
+  uint8_t len,
   volatile uint8_t *in,
   volatile uint8_t *out,
   volatile uint8_t *flags
